+++ src/chains/celo/formatters.test.ts
@@ -399,6 +399,112 @@ describe('transaction', () => {
         "value": 0n,
       }
     `)
+
+    expect(
+      transaction.format({
+        accessList: [],
+        blockHash:
+          '0x89644bbd5c8d682a2e9611170e6c1f02573d866d286f006cbf517eec7254ec2d',
+        blockNumber: '0x1',
+        chainId: '0x1',
+        feeCurrency: '0x0f16e9b0d03470827a95cdfd0cb8a8a3b46969b9',
+        from: '0xa152f8bb749c55e9943a3a0a3111d18ee2b3f94e',
+        gas: '0x2',
+        gasPrice: undefined,
+        gatewayFee: '0x3',
+        gatewayFeeRecipient: null,
+        hash: '0xa4b1f606b66105fa45cb5db23d2f6597075701e7f0e2367f4e6a39d17a8cf98b',
+        input:
+          '0x23b872dd000000000000000000000000a00f99bc38b1ecda1fd70eaa1cd31d576a9f46b0000000000000000000000000f16e9b0d03470827a95cdfd0cb8a8a3b46969b910000000000000000000000000000000000000000000000000000002b3b6fb3d0',
+        maxFeePerGas: '0x4',
+        maxPriorityFeePerGas: '0x5',
+        nonce: '0x6',
+        r: '0x5e49a7bd0534c6b6d3bbe581659424d3747f920d40ce56e48d26e5d94aac32ca',
+        s: '0x1746abe27b7c4f00bda1ec714ac1f7083e9025b6ca3b2248e439a173e4ab55e0',
+        to: '0x15d4c048f83bd7e37d49ea4c83a07267ec4203da',
+        transactionIndex: '0x7',
+        type: '0x7c',
+        v: '0x1',
+        value: '0x0',
+      }),
+    ).toMatchInlineSnapshot(`
+      {
+        "accessList": [],
+        "blockHash": "0x89644bbd5c8d682a2e9611170e6c1f02573d866d286f006cbf517eec7254ec2d",
+        "blockNumber": 1n,
+        "chainId": 1,
+        "feeCurrency": "0x0f16e9b0d03470827a95cdfd0cb8a8a3b46969b9",
+        "from": "0xa152f8bb749c55e9943a3a0a3111d18ee2b3f94e",
+        "gas": 2n,
+        "gasPrice": undefined,
+        "gatewayFee": 3n,
+        "gatewayFeeRecipient": null,
+        "hash": "0xa4b1f606b66105fa45cb5db23d2f6597075701e7f0e2367f4e6a39d17a8cf98b",
+        "input": "0x23b872dd000000000000000000000000a00f99bc38b1ecda1fd70eaa1cd31d576a9f46b0000000000000000000000000f16e9b0d03470827a95cdfd0cb8a8a3b46969b910000000000000000000000000000000000000000000000000000002b3b6fb3d0",
+        "maxFeePerGas": 4n,
+        "maxPriorityFeePerGas": 5n,
+        "nonce": 6,
+        "r": "0x5e49a7bd0534c6b6d3bbe581659424d3747f920d40ce56e48d26e5d94aac32ca",
+        "s": "0x1746abe27b7c4f00bda1ec714ac1f7083e9025b6ca3b2248e439a173e4ab55e0",
+        "to": "0x15d4c048f83bd7e37d49ea4c83a07267ec4203da",
+        "transactionIndex": 7,
+        "type": "cip42",
+        "typeHex": "0x7c",
+        "v": 1n,
+        "value": 0n,
+      }
+    `)
+
+    expect(
+      transaction.format({
+        accessList: [],
+        blockHash:
+          '0x89644bbd5c8d682a2e9611170e6c1f02573d866d286f006cbf517eec7254ec2d',
+        blockNumber: '0x1',
+        chainId: '0x1',
+        feeCurrency: '0x0f16e9b0d03470827a95cdfd0cb8a8a3b46969b9',
+        from: '0xa152f8bb749c55e9943a3a0a3111d18ee2b3f94e',
+        gas: '0x2',
+        gasPrice: undefined,
+        hash: '0xa4b1f606b66105fa45cb5db23d2f6597075701e7f0e2367f4e6a39d17a8cf98b',
+        input:
+          '0x23b872dd000000000000000000000000a00f99bc38b1ecda1fd70eaa1cd31d576a9f46b0000000000000000000000000f16e9b0d03470827a95cdfd0cb8a8a3b46969b910000000000000000000000000000000000000000000000000000002b3b6fb3d0',
+        maxFeePerGas: '0x4',
+        maxPriorityFeePerGas: '0x5',
+        nonce: '0x6',
+        r: '0x5e49a7bd0534c6b6d3bbe581659424d3747f920d40ce56e48d26e5d94aac32ca',
+        s: '0x1746abe27b7c4f00bda1ec714ac1f7083e9025b6ca3b2248e439a173e4ab55e0',
+        to: '0x15d4c048f83bd7e37d49ea4c83a07267ec4203da',
+        transactionIndex: '0x7',
+        type: '0x7b',
+        v: '0x1',
+        value: '0x0',
+      }),
+    ).toMatchInlineSnapshot(`
+      {
+        "accessList": [],
+        "blockHash": "0x89644bbd5c8d682a2e9611170e6c1f02573d866d286f006cbf517eec7254ec2d",
+        "blockNumber": 1n,
+        "chainId": 1,
+        "feeCurrency": "0x0f16e9b0d03470827a95cdfd0cb8a8a3b46969b9",
+        "from": "0xa152f8bb749c55e9943a3a0a3111d18ee2b3f94e",
+        "gas": 2n,
+        "gasPrice": undefined,
+        "hash": "0xa4b1f606b66105fa45cb5db23d2f6597075701e7f0e2367f4e6a39d17a8cf98b",
+        "input": "0x23b872dd000000000000000000000000a00f99bc38b1ecda1fd70eaa1cd31d576a9f46b0000000000000000000000000f16e9b0d03470827a95cdfd0cb8a8a3b46969b910000000000000000000000000000000000000000000000000000002b3b6fb3d0",
+        "maxFeePerGas": 4n,
+        "maxPriorityFeePerGas": 5n,
+        "nonce": 6,
+        "r": "0x5e49a7bd0534c6b6d3bbe581659424d3747f920d40ce56e48d26e5d94aac32ca",
+        "s": "0x1746abe27b7c4f00bda1ec714ac1f7083e9025b6ca3b2248e439a173e4ab55e0",
+        "to": "0x15d4c048f83bd7e37d49ea4c83a07267ec4203da",
+        "transactionIndex": 7,
+        "type": "cip64",
+        "typeHex": "0x7b",
+        "v": 1n,
+        "value": 0n,
+      }
+    `)
   })
 
   test('action', async () => {
@@ -751,9 +857,9 @@ describe('transactionReceipt', () => {
 })
 
 describe('transactionRequest', () => {
-  test('formatter', () => {
-    const { transactionRequest } = celo.formatters!
+  const { transactionRequest } = celo.formatters!
 
+  test('formatter cip42', () => {
     expect(
       transactionRequest.format({
         feeCurrency: '0x0f16e9b0d03470827a95cdfd0cb8a8a3b46969b9',
@@ -819,7 +925,6 @@ describe('transactionRequest', () => {
         maxFeePerGas: 2n,
         maxPriorityFeePerGas: 1n,
         nonce: 1,
-        type: 'cip42',
         value: 1n,
       }),
     ).toMatchInlineSnapshot(`
@@ -843,10 +948,66 @@ describe('transactionRequest', () => {
         feeCurrency: '0x0f16e9b0d03470827a95cdfd0cb8a8a3b46969b9',
         from: '0x0f16e9b0d03470827a95cdfd0cb8a8a3b46969b9',
         gas: 1n,
+        gatewayFee: 4n,
+        gatewayFeeRecipient: '0x0f16e9b0d03470827a95cdfd0cb8a8a3b46969b9',
         maxFeePerGas: 2n,
         maxPriorityFeePerGas: 1n,
         nonce: 1,
-        type: 'cip64',
+        value: 1n,
+      }),
+    ).toMatchInlineSnapshot(`
+      {
+        "feeCurrency": "0x0f16e9b0d03470827a95cdfd0cb8a8a3b46969b9",
+        "from": "0x0f16e9b0d03470827a95cdfd0cb8a8a3b46969b9",
+        "gas": "0x1",
+        "gasPrice": undefined,
+        "gatewayFee": "0x4",
+        "gatewayFeeRecipient": "0x0f16e9b0d03470827a95cdfd0cb8a8a3b46969b9",
+        "maxFeePerGas": "0x2",
+        "maxPriorityFeePerGas": "0x1",
+        "nonce": "0x1",
+        "type": "0x7c",
+        "value": "0x1",
+      }
+    `)
+
+    expect(
+      transactionRequest.format({
+        feeCurrency: '0x0f16e9b0d03470827a95cdfd0cb8a8a3b46969b9',
+        from: '0x0f16e9b0d03470827a95cdfd0cb8a8a3b46969b9',
+        gas: 1n,
+        gatewayFee: 4n,
+        gatewayFeeRecipient: '0x0f16e9b0d03470827a95cdfd0cb8a8a3b46969b9',
+        maxFeePerGas: 2n,
+        maxPriorityFeePerGas: 4n,
+        nonce: 1,
+        value: 1n,
+      }),
+    ).toMatchInlineSnapshot(`
+      {
+        "feeCurrency": "0x0f16e9b0d03470827a95cdfd0cb8a8a3b46969b9",
+        "from": "0x0f16e9b0d03470827a95cdfd0cb8a8a3b46969b9",
+        "gas": "0x1",
+        "gasPrice": undefined,
+        "gatewayFee": "0x4",
+        "gatewayFeeRecipient": "0x0f16e9b0d03470827a95cdfd0cb8a8a3b46969b9",
+        "maxFeePerGas": "0x2",
+        "maxPriorityFeePerGas": "0x4",
+        "nonce": "0x1",
+        "type": "0x7c",
+        "value": "0x1",
+      }
+    `)
+  })
+  test('formatter cip64', () => {
+    expect(
+      transactionRequest.format({
+        feeCurrency: '0x0f16e9b0d03470827a95cdfd0cb8a8a3b46969b9',
+        from: '0x0f16e9b0d03470827a95cdfd0cb8a8a3b46969b9',
+        gas: 1n,
+        maxFeePerGas: 2n,
+        maxPriorityFeePerGas: 1n,
+        nonce: 1,
         value: 1n,
       }),
     ).toMatchInlineSnapshot(`
@@ -858,6 +1019,30 @@ describe('transactionRequest', () => {
         "maxFeePerGas": "0x2",
         "maxPriorityFeePerGas": "0x1",
         "nonce": "0x1",
+        "type": "0x7b",
+        "value": "0x1",
+      }
+    `)
+
+    expect(
+      transactionRequest.format({
+        from: '0x0f16e9b0d03470827a95cdfd0cb8a8a3b46969b9',
+        to: '0x70997970c51812dc3a010c7d01b50e0d17dc79c8', // Recipient (illustrative address)
+        value: 1n,
+        feeCurrency: '0x874069Fa1Eb16D44d622F2e0Ca25eeA172369bC1', // cUSD fee currency
+        maxFeePerGas: 2n, // Special field for dynamic fee transaction type (EIP-1559)
+        maxPriorityFeePerGas: 2n, // Special field for dynamic fee transaction type (EIP-1559)
+      }),
+    ).toMatchInlineSnapshot(`
+      {
+        "feeCurrency": "0x874069Fa1Eb16D44d622F2e0Ca25eeA172369bC1",
+        "from": "0x0f16e9b0d03470827a95cdfd0cb8a8a3b46969b9",
+        "gas": undefined,
+        "gasPrice": undefined,
+        "maxFeePerGas": "0x2",
+        "maxPriorityFeePerGas": "0x2",
+        "nonce": undefined,
+        "to": "0x70997970c51812dc3a010c7d01b50e0d17dc79c8",
         "type": "0x7b",
         "value": "0x1",
       }