+++ src/chains/celo/serializers.test.ts
@@ -11,6 +11,7 @@ import {
   parseGwei,
   parseTransaction,
 } from '../../index.js'
+import { parseTransactionCelo } from './parsers.js'
 import { serializeTransactionCelo } from './serializers.js'
 import type {
   TransactionSerializableCIP42,
@@ -20,31 +21,36 @@ import type {
 const commonBaseTx = {
   to: accounts[0].address,
   chainId: 42220,
-  nonce: 0,
-  maxFeePerGas: parseGwei('2'),
-  maxPriorityFeePerGas: parseGwei('2'),
+  nonce: 1,
   feeCurrency: '0x765de816845861e75a25fca122bb6898b8b1282a',
   value: parseEther('1'),
 }
+
 const baseCip42 = {
   ...commonBaseTx,
-  type: 'cip42',
+  maxFeePerGas: parseGwei('2'),
+  maxPriorityFeePerGas: parseGwei('2'),
+  gatewayFee: 1000023434343n,
+  gatewayFeeRecipient: accounts[7].address,
 } as TransactionSerializableCIP42
 
 const baseCip64 = {
   ...commonBaseTx,
-  type: 'cip64',
+  maxFeePerGas: parseGwei('2'),
+  maxPriorityFeePerGas: parseGwei('2'),
 } as TransactionSerializableCIP64
 
 describe('cip42', () => {
   test('should be able to serialize a cip42 transaction', () => {
-    const transaction: TransactionSerializableCIP42 = {
-      ...baseCip42,
-    }
-
-    expect(serializeTransactionCelo(transaction)).toEqual(
-      '0x7cf84682a4ec80847735940084773594008094765de816845861e75a25fca122bb6898b8b1282a808094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c0',
+    // sanity checks the serialized value, but then rely on the parser
+    const serialized = serializeTransactionCelo(baseCip42)
+    const reparsed = parseTransactionCelo(serialized)
+    const reserialized = serializeTransactionCelo(reparsed)
+    expect(serialized).toEqual(
+      '0x7cf85f82a4ec01847735940084773594008094765de816845861e75a25fca122bb6898b8b1282a9414dc79964da2c08b23698b3d3cc7ca32193d995585e8d60aa46794f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c0',
     )
+    expect(reparsed).toEqual({ ...baseCip42, type: 'cip42' })
+    expect(serialized).toEqual(reserialized)
   })
 
   test('args: accessList', () => {
@@ -61,19 +67,24 @@ describe('cip42', () => {
       ],
     }
 
-    expect(serializeTransactionCelo(transaction)).toEqual(
-      '0x7cf8a282a4ec80847735940084773594008094765de816845861e75a25fca122bb6898b8b1282a808094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080f85bf859940000000000000000000000000000000000000000f842a00000000000000000000000000000000000000000000000000000000000000001a060fdd29ff912ce880cd3edaf9f932dc61d3dae823ea77e0323f94adb9f6a72fe',
+    expect(parseTransactionCelo(serializeTransactionCelo(transaction))).toEqual(
+      {
+        ...transaction,
+        type: 'cip42',
+      },
     )
   })
 
   test('args: data', () => {
-    const args: TransactionSerializableCIP42 = {
+    const transaction: TransactionSerializableCIP42 = {
       ...baseCip42,
       data: '0x1234',
     }
-    const serialized = serializeTransactionCelo(args)
-    expect(serialized).toEqual(
-      '0x7cf84882a4ec80847735940084773594008094765de816845861e75a25fca122bb6898b8b1282a808094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a7640000821234c0',
+    expect(parseTransactionCelo(serializeTransactionCelo(transaction))).toEqual(
+      {
+        ...transaction,
+        type: 'cip42',
+      },
     )
   })
 
@@ -85,8 +96,11 @@ describe('cip42', () => {
       feeCurrency: undefined,
     }
 
-    expect(serializeTransactionCelo(transaction)).toEqual(
-      '0x7cf84e82a4ec8084773594008477359400808094f39fd6e51aad88f6f4ce6ab8827279cfffb9226688016345785d8a000094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c0',
+    expect(parseTransactionCelo(serializeTransactionCelo(transaction))).toEqual(
+      {
+        ...transaction,
+        type: 'cip42',
+      },
     )
   })
 
@@ -95,8 +109,11 @@ describe('cip42', () => {
       ...baseCip42,
       gas: 69420n,
     }
-    expect(serializeTransactionCelo(transaction)).toEqual(
-      '0x7cf84982a4ec808477359400847735940083010f2c94765de816845861e75a25fca122bb6898b8b1282a808094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c0',
+    expect(parseTransactionCelo(serializeTransactionCelo(transaction))).toEqual(
+      {
+        ...transaction,
+        type: 'cip42',
+      },
     )
   })
 
@@ -105,86 +122,107 @@ describe('cip42', () => {
       ...baseCip42,
       gas: undefined,
     }
-    expect(serializeTransactionCelo(transaction)).toEqual(
-      '0x7cf84682a4ec80847735940084773594008094765de816845861e75a25fca122bb6898b8b1282a808094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c0',
+    expect(parseTransactionCelo(serializeTransactionCelo(transaction))).toEqual(
+      {
+        ...transaction,
+        type: 'cip42',
+      },
     )
   })
 
-  test('args: gatewayFee', () => {
+  test('args: gatewayFee (absent)', () => {
     const transaction: TransactionSerializableCIP42 = {
       ...baseCip42,
-      gatewayFeeRecipient: accounts[5].address,
-      gatewayFee: parseEther('0.1'),
+      gatewayFeeRecipient: undefined,
+      gatewayFee: undefined,
+      type: 'cip42',
     }
-    expect(serializeTransactionCelo(transaction)).toEqual(
-      '0x7cf86282a4ec80847735940084773594008094765de816845861e75a25fca122bb6898b8b1282a949965507d1a55bcc2695c58ba16fb37d819b0a4dc88016345785d8a000094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c0',
+    expect(parseTransactionCelo(serializeTransactionCelo(transaction))).toEqual(
+      {
+        ...transaction,
+        type: 'cip42',
+      },
     )
   })
 
   test('args: maxFeePerGas (absent)', () => {
     const transaction: TransactionSerializableCIP42 = {
       ...baseCip42,
-      // @ts-expect-error
       maxFeePerGas: undefined,
+      type: 'cip42',
     }
-    expect(serializeTransactionCelo(transaction)).toEqual(
-      '0x7cf84282a4ec808477359400808094765de816845861e75a25fca122bb6898b8b1282a808094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c0',
+    expect(parseTransactionCelo(serializeTransactionCelo(transaction))).toEqual(
+      {
+        ...transaction,
+        type: 'cip42',
+      },
     )
   })
 
   test('args: maxPriorityFeePerGas (absent)', () => {
     const transaction: TransactionSerializableCIP42 = {
       ...baseCip42,
-      // @ts-expect-error
       maxPriorityFeePerGas: undefined,
+      type: 'cip42',
     }
-    expect(serializeTransactionCelo(transaction)).toEqual(
-      '0x7cf84282a4ec808084773594008094765de816845861e75a25fca122bb6898b8b1282a808094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c0',
+    expect(parseTransactionCelo(serializeTransactionCelo(transaction))).toEqual(
+      {
+        ...transaction,
+        type: 'cip42',
+      },
     )
   })
 
-  test('args: nonce', () => {
+  test('args: nonce (absent)', () => {
     const transaction: TransactionSerializableCIP42 = {
       ...baseCip42,
-      nonce: 20,
+      nonce: undefined,
     }
-    expect(serializeTransactionCelo(transaction)).toEqual(
-      '0x7cf84682a4ec14847735940084773594008094765de816845861e75a25fca122bb6898b8b1282a808094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c0',
+    expect(parseTransactionCelo(serializeTransactionCelo(transaction))).toEqual(
+      {
+        ...transaction,
+        type: 'cip42',
+      },
     )
   })
 
   test('args: to (absent)', () => {
-    const args: TransactionSerializableCIP42 = {
+    const transaction: TransactionSerializableCIP42 = {
       ...baseCip42,
       to: undefined,
     }
-    const serialized = serializeTransactionCelo(args)
-    expect(serialized).toEqual(
-      '0x7cf282a4ec80847735940084773594008094765de816845861e75a25fca122bb6898b8b1282a808080880de0b6b3a764000080c0',
+    expect(parseTransactionCelo(serializeTransactionCelo(transaction))).toEqual(
+      {
+        ...transaction,
+        type: 'cip42',
+      },
     )
   })
 
   test('args: value (absent)', () => {
-    const args: TransactionSerializableCIP42 = {
+    const transaction: TransactionSerializableCIP42 = {
       ...baseCip42,
       value: undefined,
     }
-    const serialized = serializeTransactionCelo(args)
-    expect(serialized).toEqual(
-      '0x7cf83e82a4ec80847735940084773594008094765de816845861e75a25fca122bb6898b8b1282a808094f39fd6e51aad88f6f4ce6ab8827279cfffb922668080c0',
+    expect(parseTransactionCelo(serializeTransactionCelo(transaction))).toEqual(
+      {
+        ...transaction,
+        type: 'cip42',
+      },
     )
   })
 
   test('type is undefined but has cip42 fields', () => {
     const transaction: TransactionSerializableCIP42 = {
       ...baseCip42,
-      gatewayFeeRecipient: accounts[7].address,
-      gatewayFee: 1000023434343n,
       type: undefined,
     }
 
-    expect(serializeTransactionCelo(transaction)).toEqual(
-      '0x7cf85f82a4ec80847735940084773594008094765de816845861e75a25fca122bb6898b8b1282a9414dc79964da2c08b23698b3d3cc7ca32193d995585e8d60aa46794f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c0',
+    expect(parseTransactionCelo(serializeTransactionCelo(transaction))).toEqual(
+      {
+        ...transaction,
+        type: 'cip42',
+      },
     )
   })
 
@@ -195,12 +233,20 @@ describe('cip42', () => {
       serializer: serializeTransactionCelo,
     })
 
-    expect(signed).toEqual(
-      '0x7cf88982a4ec80847735940084773594008094765de816845861e75a25fca122bb6898b8b1282a808094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c080a01ae1d60446ad5fdd620e1982050dc315ff9a0f61b32bcc2a3cdadd0571a76df7a073aba459b3aef6796d5f2a9979551c29f66586821b5613d5080d00782b07c280',
-    )
+    const serialized =
+      '0x7cf8a282a4ec01847735940084773594008094765de816845861e75a25fca122bb6898b8b1282a9414dc79964da2c08b23698b3d3cc7ca32193d995585e8d60aa46794f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c001a0430169754e015d53ccf07816d580ea968cbf1ccec6fcf45589507c7a38c0e685a00a66965256aee1476446f293f6b7db29ddcfe09eba707612d63e3b67035671fe'
+    expect(signed).toEqual(serialized)
+    expect(parseTransactionCelo(signed)).toEqual({
+      ...baseCip42,
+      type: 'cip42',
+    })
   })
 
   test('signature', () => {
+    const tx1 =
+      '0x7cf8a282a4ec01847735940084773594008094765de816845861e75a25fca122bb6898b8b1282a9414dc79964da2c08b23698b3d3cc7ca32193d995585e8d60aa46794f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c001a060fdd29ff912ce880cd3edaf9f932dc61d3dae823ea77e0323f94adb9f6a72fea060fdd29ff912ce880cd3edaf9f932dc61d3dae823ea77e0323f94adb9f6a72fe'
+    const tx2 =
+      '0x7cf8a282a4ec01847735940084773594008094765de816845861e75a25fca122bb6898b8b1282a9414dc79964da2c08b23698b3d3cc7ca32193d995585e8d60aa46794f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c080a060fdd29ff912ce880cd3edaf9f932dc61d3dae823ea77e0323f94adb9f6a72fea060fdd29ff912ce880cd3edaf9f932dc61d3dae823ea77e0323f94adb9f6a72fe'
     expect(
       serializeTransactionCelo(
         baseCip42,
@@ -211,9 +257,7 @@ describe('cip42', () => {
           v: 28n,
         },
       ),
-    ).toEqual(
-      '0x7cf88982a4ec80847735940084773594008094765de816845861e75a25fca122bb6898b8b1282a808094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c001a060fdd29ff912ce880cd3edaf9f932dc61d3dae823ea77e0323f94adb9f6a72fea060fdd29ff912ce880cd3edaf9f932dc61d3dae823ea77e0323f94adb9f6a72fe',
-    )
+    ).toEqual(tx1)
     expect(
       serializeTransactionCelo(
         baseCip42,
@@ -224,21 +268,24 @@ describe('cip42', () => {
           v: 27n,
         },
       ),
-    ).toEqual(
-      '0x7cf88982a4ec80847735940084773594008094765de816845861e75a25fca122bb6898b8b1282a808094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c080a060fdd29ff912ce880cd3edaf9f932dc61d3dae823ea77e0323f94adb9f6a72fea060fdd29ff912ce880cd3edaf9f932dc61d3dae823ea77e0323f94adb9f6a72fe',
-    )
+    ).toEqual(tx2)
+
+    expect(parseTransactionCelo(tx1)).toEqual(parseTransactionCelo(tx2))
+    expect(parseTransactionCelo(tx1)).toEqual({ ...baseCip42, type: 'cip42' })
   })
 })
 
 describe('cip64', () => {
   test('should be able to serialize a cip64 transaction', () => {
-    const transaction: TransactionSerializableCIP64 = {
-      ...baseCip64,
-    }
-
-    expect(serializeTransactionCelo(transaction)).toEqual(
-      '0x7bf84482a4ec80847735940084773594008094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c094765de816845861e75a25fca122bb6898b8b1282a',
+    // sanity checks the serialized value, but then rely on the parser
+    const serialized = serializeTransactionCelo(baseCip64)
+    const reparsed = parseTransactionCelo(serialized)
+    const reserialized = serializeTransactionCelo(reparsed)
+    expect(serialized).toEqual(
+      '0x7bf84482a4ec01847735940084773594008094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c094765de816845861e75a25fca122bb6898b8b1282a',
     )
+    expect(reparsed).toEqual({ ...baseCip64, type: 'cip64' })
+    expect(serialized).toEqual(reserialized)
   })
 
   test('args: accessList', () => {
@@ -255,19 +302,24 @@ describe('cip64', () => {
       ],
     }
 
-    expect(serializeTransactionCelo(transaction)).toEqual(
-      '0x7bf8a082a4ec80847735940084773594008094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080f85bf859940000000000000000000000000000000000000000f842a00000000000000000000000000000000000000000000000000000000000000001a060fdd29ff912ce880cd3edaf9f932dc61d3dae823ea77e0323f94adb9f6a72fe94765de816845861e75a25fca122bb6898b8b1282a',
+    expect(parseTransactionCelo(serializeTransactionCelo(transaction))).toEqual(
+      {
+        ...transaction,
+        type: 'cip64',
+      },
     )
   })
 
   test('args: data', () => {
-    const args: TransactionSerializableCIP64 = {
+    const transaction: TransactionSerializableCIP64 = {
       ...baseCip64,
       data: '0x1234',
     }
-    const serialized = serializeTransactionCelo(args)
-    expect(serialized).toEqual(
-      '0x7bf84682a4ec80847735940084773594008094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a7640000821234c094765de816845861e75a25fca122bb6898b8b1282a',
+    expect(parseTransactionCelo(serializeTransactionCelo(transaction))).toEqual(
+      {
+        ...transaction,
+        type: 'cip64',
+      },
     )
   })
 
@@ -276,8 +328,11 @@ describe('cip64', () => {
       ...baseCip64,
       gas: 69420n,
     }
-    expect(serializeTransactionCelo(transaction)).toEqual(
-      '0x7bf84782a4ec808477359400847735940083010f2c94f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c094765de816845861e75a25fca122bb6898b8b1282a',
+    expect(parseTransactionCelo(serializeTransactionCelo(transaction))).toEqual(
+      {
+        ...transaction,
+        type: 'cip64',
+      },
     )
   })
 
@@ -286,62 +341,78 @@ describe('cip64', () => {
       ...baseCip64,
       gas: undefined,
     }
-    expect(serializeTransactionCelo(transaction)).toEqual(
-      '0x7bf84482a4ec80847735940084773594008094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c094765de816845861e75a25fca122bb6898b8b1282a',
+    expect(parseTransactionCelo(serializeTransactionCelo(transaction))).toEqual(
+      {
+        ...transaction,
+        type: 'cip64',
+      },
     )
   })
 
   test('args: maxFeePerGas (absent)', () => {
     const transaction: TransactionSerializableCIP64 = {
       ...baseCip64,
-      // @ts-expect-error
+      type: 'cip64',
       maxFeePerGas: undefined,
     }
-    expect(serializeTransactionCelo(transaction)).toEqual(
-      '0x7bf84082a4ec808477359400808094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c094765de816845861e75a25fca122bb6898b8b1282a',
+    expect(parseTransactionCelo(serializeTransactionCelo(transaction))).toEqual(
+      {
+        ...transaction,
+        type: 'cip64',
+      },
     )
   })
 
   test('args: maxPriorityFeePerGas (absent)', () => {
     const transaction: TransactionSerializableCIP64 = {
       ...baseCip64,
-      // @ts-expect-error
+      type: 'cip64',
       maxPriorityFeePerGas: undefined,
     }
-    expect(serializeTransactionCelo(transaction)).toEqual(
-      '0x7bf84082a4ec808084773594008094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c094765de816845861e75a25fca122bb6898b8b1282a',
+    expect(parseTransactionCelo(serializeTransactionCelo(transaction))).toEqual(
+      {
+        ...transaction,
+        type: 'cip64',
+      },
     )
   })
 
-  test('args: nonce', () => {
+  test('args: nonce (absent)', () => {
     const transaction: TransactionSerializableCIP64 = {
       ...baseCip64,
-      nonce: 20,
+      nonce: undefined,
     }
-    expect(serializeTransactionCelo(transaction)).toEqual(
-      '0x7bf84482a4ec14847735940084773594008094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c094765de816845861e75a25fca122bb6898b8b1282a',
+    expect(parseTransactionCelo(serializeTransactionCelo(transaction))).toEqual(
+      {
+        ...transaction,
+        type: 'cip64',
+      },
     )
   })
 
   test('args: to (absent)', () => {
-    const args: TransactionSerializableCIP64 = {
+    const transaction: TransactionSerializableCIP64 = {
       ...baseCip64,
       to: undefined,
     }
-    const serialized = serializeTransactionCelo(args)
-    expect(serialized).toEqual(
-      '0x7bf082a4ec80847735940084773594008080880de0b6b3a764000080c094765de816845861e75a25fca122bb6898b8b1282a',
+    expect(parseTransactionCelo(serializeTransactionCelo(transaction))).toEqual(
+      {
+        ...transaction,
+        type: 'cip64',
+      },
     )
   })
 
   test('args: value (absent)', () => {
-    const args: TransactionSerializableCIP64 = {
+    const transaction: TransactionSerializableCIP64 = {
       ...baseCip64,
       value: undefined,
     }
-    const serialized = serializeTransactionCelo(args)
-    expect(serialized).toEqual(
-      '0x7bf83c82a4ec80847735940084773594008094f39fd6e51aad88f6f4ce6ab8827279cfffb922668080c094765de816845861e75a25fca122bb6898b8b1282a',
+    expect(parseTransactionCelo(serializeTransactionCelo(transaction))).toEqual(
+      {
+        ...transaction,
+        type: 'cip64',
+      },
     )
   })
 
@@ -351,9 +422,11 @@ describe('cip64', () => {
       feeCurrency: '0xd8763cba276a3738e6de85b4b3bf5fded6d6ca73',
       type: undefined,
     }
-
-    expect(serializeTransactionCelo(transaction)).toEqual(
-      '0x7bf84482a4ec80847735940084773594008094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c094d8763cba276a3738e6de85b4b3bf5fded6d6ca73',
+    expect(parseTransactionCelo(serializeTransactionCelo(transaction))).toEqual(
+      {
+        ...transaction,
+        type: 'cip64',
+      },
     )
   })
 
@@ -363,13 +436,20 @@ describe('cip64', () => {
       transaction: baseCip64,
       serializer: serializeTransactionCelo,
     })
-
-    expect(signed).toEqual(
-      '0x7bf88782a4ec80847735940084773594008094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c094765de816845861e75a25fca122bb6898b8b1282a80a07b5ef5199c55a6765782eeb966fe135ff34b39eadabf952dfc00b017924b356aa06425ed31cf71b817c064b669f89d819ee25affa1669270b9b8ac9638b53d7e7f',
-    )
+    const serialized =
+      '0x7bf88782a4ec01847735940084773594008094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c094765de816845861e75a25fca122bb6898b8b1282a01a087a588ffb7bec68b00d264932305b3badc3bfba2f48e9d3d916d038e1bb831a5a077e665b35849c636b52b3cc205acb98141b4a582fb84a7181e048dc5473c6d6d'
+    expect(signed).toEqual(serialized)
+    expect(parseTransactionCelo(signed)).toEqual({
+      ...baseCip64,
+      type: 'cip64',
+    })
   })
 
   test('signature', () => {
+    const tx1 =
+      '0x7bf88782a4ec01847735940084773594008094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c094765de816845861e75a25fca122bb6898b8b1282a01a060fdd29ff912ce880cd3edaf9f932dc61d3dae823ea77e0323f94adb9f6a72fea060fdd29ff912ce880cd3edaf9f932dc61d3dae823ea77e0323f94adb9f6a72fe'
+    const tx2 =
+      '0x7bf88782a4ec01847735940084773594008094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c094765de816845861e75a25fca122bb6898b8b1282a80a060fdd29ff912ce880cd3edaf9f932dc61d3dae823ea77e0323f94adb9f6a72fea060fdd29ff912ce880cd3edaf9f932dc61d3dae823ea77e0323f94adb9f6a72fe'
     expect(
       serializeTransactionCelo(
         baseCip64,
@@ -380,9 +460,7 @@ describe('cip64', () => {
           v: 28n,
         },
       ),
-    ).toEqual(
-      '0x7bf88782a4ec80847735940084773594008094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c094765de816845861e75a25fca122bb6898b8b1282a01a060fdd29ff912ce880cd3edaf9f932dc61d3dae823ea77e0323f94adb9f6a72fea060fdd29ff912ce880cd3edaf9f932dc61d3dae823ea77e0323f94adb9f6a72fe',
-    )
+    ).toEqual(tx1)
     expect(
       serializeTransactionCelo(
         baseCip64,
@@ -393,16 +471,20 @@ describe('cip64', () => {
           v: 27n,
         },
       ),
-    ).toEqual(
-      '0x7bf88782a4ec80847735940084773594008094f39fd6e51aad88f6f4ce6ab8827279cfffb92266880de0b6b3a764000080c094765de816845861e75a25fca122bb6898b8b1282a80a060fdd29ff912ce880cd3edaf9f932dc61d3dae823ea77e0323f94adb9f6a72fea060fdd29ff912ce880cd3edaf9f932dc61d3dae823ea77e0323f94adb9f6a72fe',
-    )
+    ).toEqual(tx2)
+    expect(parseTransactionCelo(tx1)).toEqual(parseTransactionCelo(tx2))
+    expect(parseTransactionCelo(tx1)).toEqual({ ...baseCip64, type: 'cip64' })
   })
 })
 
 describe('invalid params specific to CIP-42', () => {
+  const baseCip42WithType = {
+    ...baseCip42,
+    type: 'cip42',
+  } as const
   test('only one of the gateWayFee fields is defined', () => {
     const transactionA: TransactionSerializableCIP42 = {
-      ...baseCip42,
+      ...baseCip42WithType,
       gatewayFee: undefined,
       gatewayFeeRecipient: accounts[7].address,
     }
@@ -410,7 +492,7 @@ describe('invalid params specific to CIP-42', () => {
       '`gatewayFee` and `gatewayFeeRecipient` must be provided together.',
     )
     const transactionB: TransactionSerializableCIP42 = {
-      ...baseCip42,
+      ...baseCip42WithType,
       gatewayFee: 1000023434343n,
       gatewayFeeRecipient: undefined,
     }
@@ -420,7 +502,7 @@ describe('invalid params specific to CIP-42', () => {
   })
   test('transaction looks like cip42 but does not have values for either feeCurrency or gatewayFee', () => {
     const transaction: TransactionSerializableCIP42 = {
-      ...baseCip42,
+      ...baseCip42WithType,
       feeCurrency: undefined,
       gatewayFee: undefined,
       gatewayFeeRecipient: undefined,
@@ -435,6 +517,7 @@ describe('invalid params specific to CIP-64', () => {
   test('transaction looks like cip64 but does not have a value for feeCurrency', () => {
     const transaction: TransactionSerializableCIP64 = {
       ...baseCip64,
+      type: 'cip64',
       feeCurrency: undefined,
     }
     expect(() => serializeTransactionCelo(transaction)).toThrowError(
@@ -443,15 +526,90 @@ describe('invalid params specific to CIP-64', () => {
   })
 })
 
-describe.each([
-  { typeName: 'CIP-42', baseTransaction: baseCip42 },
-  { typeName: 'CIP-64', baseTransaction: baseCip64 },
-])('Common invalid params (for $typeName)', ({ typeName, baseTransaction }) => {
+describe('Common invalid params (for CIP-42)', () => {
   test('invalid to', () => {
-    const transaction:
-      | TransactionSerializableCIP42
-      | TransactionSerializableCIP64 = {
-      ...baseTransaction,
+    const transaction: TransactionSerializableCIP42 = {
+      ...baseCip42,
+      to: '0xdeadbeef',
+    }
+    expect(() => serializeTransactionCelo(transaction)).toThrowError(
+      InvalidAddressError,
+    )
+  })
+
+  test('gatewayFeeRecipient is not an address', () => {
+    const transaction: TransactionSerializableCIP42 = {
+      ...baseCip42,
+      type: 'cip42',
+      // @ts-expect-error - (Type '"example"' is not assignable to type "`0x${string}"'
+      gatewayFeeRecipient: 'example',
+    }
+    expect(() => serializeTransactionCelo(transaction)).toThrowError(
+      InvalidAddressError,
+    )
+  })
+
+  test('maxFeePerGas is too high', () => {
+    const transaction: TransactionSerializableCIP42 = {
+      ...baseCip42,
+      maxFeePerGas: 2n ** 257n,
+    }
+    expect(() => serializeTransactionCelo(transaction)).toThrowError(
+      FeeCapTooHighError,
+    )
+  })
+
+  test('feeCurrency is not an address', () => {
+    const transaction: TransactionSerializableCIP42 = {
+      ...baseCip42,
+      // @ts-expect-error - (Type '"CUSD"' is not assignable to type "`0x${string}"'
+      feeCurrency: 'CUSD',
+    }
+
+    expect(() => serializeTransactionCelo(transaction)).toThrowError(
+      '`feeCurrency` MUST be a token address for CIP-42 transactions.',
+    )
+  })
+
+  test('gasPrice is defined', () => {
+    const transaction: TransactionSerializableCIP42 = {
+      ...baseCip42,
+      // @ts-expect-error
+      gasPrice: BigInt(1),
+    }
+
+    expect(() => serializeTransactionCelo(transaction)).toThrowError(
+      '`gasPrice` is not a valid CIP-42 Transaction attribute.',
+    )
+  })
+
+  test('chainID is invalid', () => {
+    const transaction: TransactionSerializableCIP42 = {
+      ...baseCip42,
+      chainId: -1,
+    }
+
+    expect(() => serializeTransactionCelo(transaction)).toThrowError(
+      `Chain ID "${-1}" is invalid.`,
+    )
+  })
+
+  test('maxPriorityFeePerGas is higher than maxPriorityFee', () => {
+    const transaction: TransactionSerializableCIP42 = {
+      ...baseCip42,
+      maxPriorityFeePerGas: parseGwei('5000000000'),
+      maxFeePerGas: parseGwei('1'),
+    }
+    expect(() => serializeTransactionCelo(transaction)).toThrowError(
+      TipAboveFeeCapError,
+    )
+  })
+})
+
+describe('Common invalid params (for CIP-64)', () => {
+  test('invalid to', () => {
+    const transaction: TransactionSerializableCIP64 = {
+      ...baseCip64,
       to: '0xdeadbeef',
     }
     expect(() => serializeTransactionCelo(transaction)).toThrowError(
@@ -460,10 +618,8 @@ describe.each([
   })
 
   test('maxPriorityFeePerGas is higher than maxPriorityFee', () => {
-    const transaction:
-      | TransactionSerializableCIP42
-      | TransactionSerializableCIP64 = {
-      ...baseTransaction,
+    const transaction: TransactionSerializableCIP64 = {
+      ...baseCip64,
       maxPriorityFeePerGas: parseGwei('5000000000'),
       maxFeePerGas: parseGwei('1'),
     }
@@ -473,10 +629,8 @@ describe.each([
   })
 
   test('maxFeePerGas is too high', () => {
-    const transaction:
-      | TransactionSerializableCIP42
-      | TransactionSerializableCIP64 = {
-      ...baseTransaction,
+    const transaction: TransactionSerializableCIP64 = {
+      ...baseCip64,
       maxPriorityFeePerGas: parseGwei('5000000000'),
       maxFeePerGas:
         115792089237316195423570985008687907853269984665640564039457584007913129639938n,
@@ -487,38 +641,32 @@ describe.each([
   })
 
   test('feeCurrency is not an address', () => {
-    const transaction:
-      | TransactionSerializableCIP42
-      | TransactionSerializableCIP64 = {
-      ...baseTransaction,
+    const transaction: TransactionSerializableCIP64 = {
+      ...baseCip64,
       // @ts-expect-error
       feeCurrency: 'CUSD',
     }
 
     expect(() => serializeTransactionCelo(transaction)).toThrowError(
-      `\`feeCurrency\` MUST be a token address for ${typeName} transactions.`,
+      '`feeCurrency` MUST be a token address for CIP-64 transactions.',
     )
   })
 
   test('gasPrice is defined', () => {
-    const transaction:
-      | TransactionSerializableCIP42
-      | TransactionSerializableCIP64 = {
-      ...baseTransaction,
+    const transaction: TransactionSerializableCIP64 = {
+      ...baseCip64,
       // @ts-expect-error
       gasPrice: BigInt(1),
     }
 
     expect(() => serializeTransactionCelo(transaction)).toThrowError(
-      `\`gasPrice\` is not a valid ${typeName} Transaction attribute.`,
+      '`gasPrice` is not a valid CIP-64 Transaction attribute.',
     )
   })
 
   test('chainID is invalid', () => {
-    const transaction:
-      | TransactionSerializableCIP42
-      | TransactionSerializableCIP64 = {
-      ...baseTransaction,
+    const transaction: TransactionSerializableCIP64 = {
+      ...baseCip64,
       chainId: -1,
     }