+++ src/chains/definitions/gnosis.ts
@@ -9,7 +9,10 @@ export const gnosis = /*#__PURE__*/ defineChain({
     symbol: 'xDAI',
   },
   rpcUrls: {
-    default: { http: ['https://rpc.gnosischain.com'] },
+    default: {
+      http: ['https://rpc.gnosischain.com'],
+      webSocket: ['wss://rpc.gnosischain.com/wss'],
+    },
   },
   blockExplorers: {
     default: {