+++ src/chains/definitions/scrollSepolia.ts
@@ -7,7 +7,6 @@ export const scrollSepolia = /*#__PURE__*/ defineChain({
   rpcUrls: {
     default: {
       http: ['https://sepolia-rpc.scroll.io'],
-      webSocket: ['wss://sepolia-rpc.scroll.io/ws'],
     },
   },
   blockExplorers: {